переписывание проекта на MVC(пока не работает)
